requirements.txt
pytermgui>=7.2.0
pyaml>=21.10.1
pycryptodome>=3.15.0